Consolidate enhanced configuration into config.py and remove separate config file
